# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Rooms API
  version: 0.1.0
  description: |
    Draft contract for the Rooms MVP backend. The current GitHub Pages build is
    still a frontend prototype and does not serve these endpoints.
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.rooms.example
    description: Production placeholder
...
paths:
...
  /v1/cities/{cityId}/stats:
    get:
      tags: [Public]
      operationId: getCityStats
      parameters:
        - name: cityId
          in: path
          required: true
          description: City id returned by /v1/cities or its exact display name.
          schema: {type: string}
      responses:
        '200':
          description: Public supply counts and privacy-safe audience bucket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CityStats'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/rooms:
    get:
      tags: [Public]
      operationId: searchRooms
      parameters:
        - {name: city, in: query, required: true, schema: {type: string}}
        - {name: date, in: query, schema: {type: string, format: date}}
        - {name: time, in: query, schema: {type: string, pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'}}
        - {name: durationMinutes, in: query, schema: {type: integer, minimum: 30, multipleOf: 30}}
        - {name: guests, in: query, schema: {type: integer, minimum: 1}}
        - {name: type, in: query, schema: {type: string}}
        - {name: features, in: query, schema: {type: array, items: {type: string}}, style: form, explode: false}
        - {name: maxPricePerHour, in: query, schema: {type: number, minimum: 0}}
        - {name: sort, in: query, schema: {type: string, enum: [rating, price, capacity], default: rating}}
        - {name: cursor, in: query, schema: {type: string}}
      responses:
        '200':
          description: Published rooms and optional availability summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomSearchResult'
        '400':
          $ref: '#/components/responses/ValidationError'
  /v1/rooms/{roomId}:
    get:
      tags: [Public]
      operationId: getRoom
      parameters:
        - $ref: '#/components/parameters/RoomId'
        - {name: date, in: query, schema: {type: string, format: date}}
      responses:
        '200':
          description: Room detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/rooms/{roomId}/reviews:
    get:
      tags: [Public]
      operationId: listRoomReviews
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '200':
          description: Approved reviews for a published room with privacy-safe author names
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicReviewPage'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/availability/search:
    post:
      tags: [Public]
      operationId: searchAvailability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityRequest'
      responses:
        '200':
          description: Free starts and maximum duration for each start
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RoomId:
      name: roomId
      in: path
      required: true
      description: Permanent UUID or a legacy public slug during frontend migration.
      schema: {type: string}
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
    NotFound:
      description: Entity not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
    City:
      type: object
      required: [id, name, active]
      properties:
        id: {type: string}
        name: {type: string}
        active: {type: boolean}
        pilot: {type: boolean}
    CityStats:
      type: object
      required: [city, publishedVenues, publishedRooms, activeClientsLabel, audienceStage, updatedAt]
      properties:
        city: {$ref: '#/components/schemas/City'}
        publishedVenues: {type: integer, minimum: 0}
        publishedRooms: {type: integer, minimum: 0}
        activeClientsLabel:
          type: [string, 'null']
          description: Public threshold such as 20+, 50+ or 100+; null while the city is launching.
        audienceStage: {type: string, enum: [launching, growing, established]}
        updatedAt: {type: string, format: date-time}
...
    Venue:
      type: object
      required: [id, slug, title, city, address, publicationStatus, partnerMode]
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        title: {type: string}
        city: {type: string}
        address: {type: string}
        latitude: {type: [number, 'null']}
        longitude: {type: [number, 'null']}
        description: {type: [string, 'null']}
        rules: {type: [string, 'null']}
        amenities: {type: array, items: {type: string}}
        publicationStatus: {type: string, enum: [review, published, hidden]}
        partnerMode: {type: string, enum: [catalog, crm]}
        paymentMethods: {type: array, items: {type: string, enum: [card, cash]}}
    RoomSummary:
      type: object
      required: [id, slug, venue, title, subtitle, type, capacityMax, pricePerHour, minimumHours, photos]
      properties:
        id: {type: string, format: uuid}
        slug: {type: string}
        venue: {$ref: '#/components/schemas/Venue'}
        title: {type: string}
        subtitle: {type: string}
        type: {type: string}
        capacityMin: {type: integer, minimum: 1}
        capacityMax: {type: integer, minimum: 1}
        pricePerHour: {type: number, minimum: 0}
        minimumHours: {type: number, minimum: 0.5, multipleOf: 0.5}
        rating: {type: number, minimum: 0, maximum: 5}
        reviewCount: {type: integer, minimum: 0}
        features: {type: array, items: {type: string}}
        tags: {type: array, items: {type: string}}
        promotion: {type: [string, 'null']}
        photos: {type: array, items: {type: string, format: uri}}
        nearestWindows: {type: array, items: {$ref: '#/components/schemas/AvailabilityWindow'}}
    RoomDetail:
      allOf:
        - $ref: '#/components/schemas/RoomSummary'
        - type: object
          properties:
            description: {type: string}
            rules: {type: string}
            opensAtHour: {type: number, minimum: 0, maximum: 47.5, multipleOf: 0.5}
            closesAtHour: {type: number, minimum: 0, maximum: 48, multipleOf: 0.5}
            bufferMinutes: {type: integer, enum: [0, 15, 30, 45, 60]}
            services: {type: array, items: {$ref: '#/components/schemas/RoomService'}}
            availability: {$ref: '#/components/schemas/AvailabilityResponse'}
...
    RoomService:
      type: object
      required: [id, name, price]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        description: {type: [string, 'null']}
        price: {type: number, minimum: 0}
    RoomSearchResult:
      type: object
      required: [items, hasMore]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/RoomSummary'}}
        nextCursor: {type: [string, 'null']}
        hasMore: {type: boolean}
    PublicReview:
      type: object
      required: [id, roomId, authorName, rating, publishedAt]
      properties:
        id: {type: string, format: uuid}
        roomId: {type: string, format: uuid}
        authorName: {type: string, description: First name or a generic privacy-safe label}
        rating: {type: integer, minimum: 1, maximum: 5}
        body: {type: [string, 'null']}
        partnerReply: {type: [string, 'null']}
        publishedAt: {type: string, format: date-time}
    PublicReviewPage:
      type: object
      required: [items, hasMore]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/PublicReview'}}
        nextCursor: {type: [string, 'null']}
        hasMore: {type: boolean, const: false}
    AvailabilityRequest:
      type: object
      required: [roomIds, date, durationMinutes]
      properties:
        roomIds:
          type: array
          minItems: 1
          items: {type: string, format: uuid}
        date: {type: string, format: date}
        durationMinutes: {type: integer, minimum: 30, multipleOf: 30}
        preferredTime: {type: [string, 'null']}
        guests: {type: integer, minimum: 1}
    AvailabilityResponse:
      type: object
      required: [date, timezone, windows]
      properties:
        date: {type: string, format: date}
        timezone: {type: string, example: Europe/Moscow}
        windows: {type: array, items: {$ref: '#/components/schemas/AvailabilityWindow'}}
    AvailabilityWindow:
      type: object
      required: [startsAt, maximumDurationMinutes]
      properties:
        startsAt: {type: string, format: date-time}
        maximumDurationMinutes: {type: integer, minimum: 30, multipleOf: 30}
        exactMatch: {type: boolean}
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: {type: string}
        message: {type: string}
        requestId: {type: string}
        fields:
          type: object
          additionalProperties: {type: string}